+++ tcpb/clients.py
@@ -17,18 +17,25 @@ import socket
 import struct
 import warnings
 from time import sleep
-from typing import Any, Dict, List, Optional
+from typing import Any, Dict, List, Optional, Union
 from uuid import uuid4
 
 import httpx
 import numpy as np
+from google.protobuf.json_format import MessageToDict
 from qcelemental.models.results import (
     AtomicInput,
     AtomicResult,
     NativeFilesProtocolEnum,
+    Provenance,
 )
 
-from tcpb.utils import atomic_input_to_job_input, job_output_to_atomic_result
+from tcpb.utils import (  # job_output_to_atomic_result,
+    SUPPORTED_DRIVERS,
+    atomic_input_to_job_input,
+    to_atomic_result_properties,
+    to_wavefunction_properties,
+)
 
 # Import the Protobuf messages generated from the .proto file
 from . import terachem_server_pb2 as pb
@@ -43,6 +50,8 @@ class TCProtobufClient:
     """Connect and communicate with a TeraChem instance running in Protocol Buffer server mode
     (i.e. TeraChem was started with the -s|--server flag)
     """
+
+    creator = "terachem_pbs"
 
     def __init__(
         self, host: str = "127.0.0.1", port: int = 11111, debug=False, trace=False
@@ -156,9 +165,84 @@ class TCProtobufClient:
             sleep(0.5)
 
         job_output = self._recv_msg(pb.JOBOUTPUT)
-        return job_output_to_atomic_result(
+        return self.job_output_to_atomic_result(
             atomic_input=atomic_input, job_output=job_output
         )
+
+    def job_output_to_atomic_result(
+        self,
+        *,
+        atomic_input: AtomicInput,
+        job_output: pb.JobOutput,
+    ) -> AtomicResult:
+        """Convert JobOutput to AtomicResult"""
+        # Convert job_output to python types
+        # NOTE: Required so that AtomicResult is JSON serializable. Protobuf types are not.
+        jo_dict = MessageToDict(job_output, preserving_proto_field_name=True)
+
+        if atomic_input.driver.upper() == "ENERGY":
+            # Select first element in list (ground state); may need to modify for excited
+            # states
+            return_result: Union[float, List[float]] = jo_dict["energy"][0]
+
+        elif atomic_input.driver.upper() == "GRADIENT":
+            return_result = jo_dict["gradient"]
+
+        else:
+            raise ValueError(
+                f"Unsupported driver: {atomic_input.driver.upper()}, supported drivers "
+                f"include: {SUPPORTED_DRIVERS}"
+            )
+
+        # Prepare AtomicInput to be base input for AtomicResult
+        atomic_input_dict = atomic_input.dict()
+        atomic_input_dict.pop("provenance", None)
+
+        # Create AtomicResult as superset of AtomicInput values
+        atomic_result = AtomicResult(
+            **atomic_input_dict,
+            # Create new provenance object
+            provenance=Provenance(
+                creator=self.creator,
+                version="",
+                routine=f"tcpb.{self.__class__.__name__}.compute",
+            ),
+            return_result=return_result,
+            properties=to_atomic_result_properties(job_output),
+            # NOTE: Wavefunction will only be added if atomic_input.protocols.wavefunction != 'none'
+            wavefunction=to_wavefunction_properties(job_output, atomic_input),
+            success=True,
+        )
+        # And extend extras to include values additional to input extras
+        atomic_result.extras.update(
+            {
+                settings.extras_qcvars_kwarg: {
+                    "charges": jo_dict.get("charges"),
+                    "spins": jo_dict.get("spins"),
+                    "meyer_bond_order": jo_dict.get("bond_order"),
+                    "orb_size": jo_dict.get("orb_size"),
+                    "excited_state_energies": jo_dict.get("energy"),
+                    "cis_transition_dipoles": jo_dict.get("cis_transition_dipoles"),
+                    "compressed_bond_order": jo_dict.get("compressed_bond_order"),
+                    "compressed_hessian": jo_dict.get("compressed_hessian"),
+                    "compressed_ao_data": jo_dict.get("compressed_ao_data"),
+                    "compressed_primitive_data": jo_dict.get(
+                        "compressed_primitive_data"
+                    ),
+                    "compressed_mo_vector": jo_dict.get("compressed_mo_vector"),
+                    "imd_mmatom_gradient": jo_dict.get("imd_mmatom_gradient"),
+                },
+                settings.extras_job_kwarg: {
+                    "job_dir": jo_dict.get("job_dir"),
+                    "job_scr_dir": jo_dict.get("job_scr_dir"),
+                    "server_job_id": jo_dict.get("server_job_id"),
+                    "orb1afile": jo_dict.get("orb1afile"),
+                    "orb1bfile": jo_dict.get("orb1bfile"),
+                },
+            }
+        )
+
+        return atomic_result
 
     def send_job_async(self, jobType="energy", geom=None, unitType="bohr", **kwargs):
         """Pack and send the current JobInput to the TeraChem Protobuf server asynchronously.
@@ -210,7 +294,7 @@ class TCProtobufClient:
         """Sets status on self if job is accepted"""
         warnings.warn(
             "The status returned from the TCPB Server is off by one. The status "
-            "returned contains actually contains values for the previous job."
+            "returned actually contains values for the previous job."
         )
         self.curr_job_dir = status_msg.job_dir
         self.curr_job_scr_dir = status_msg.job_scr_dir
@@ -851,6 +935,8 @@ class TCFrontEndClient(TCProtobufClient):
     be put to the server e.g., to use as an initial wave function guess, or any output
     file retrieved after a computation.
     """
+
+    creator = "terachem_fe"
 
     def __init__(
         self,